services/ai-agent-service/pyproject.toml
[project]
name = "ecoclear-ai-agent"
version = "1.0.0"
description = "LangGraph multi-agent supervisor for CECB environmental clearance scrutiny"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langsmith>=0.1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
]
